requirement.txt
amqp==1.4.9
anyjson==0.3.3
Babel==2.8.0
bcrypt==3.1.7
billiard==3.3.0.23
celery==3.1.26.post2
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
colorlog==4.1.0
cryptography==2.8
cycler==0.10.0
Django==2.0.3
django-celery==3.2.2
docopt==0.6.2
dwebsocket==0.4.2
eventlet==0.22.1
fake-useragent==0.1.11
flower==0.9.2
greenlet==0.4.15
har2case==0.3.1
HttpRunner==1.5.8
idna==2.6
Jinja2==2.11.1
kiwisolver==1.2.0
kombu==3.0.37
MarkupSafe==1.1.1
matplotlib==3.0.3
module==0.0.4
mysql==0.0.2
mysqlclient==1.3.12
nose==1.3.7
nose-html-reporting==0.2.3
numpy==1.18.1
paramiko==2.4.1
Pillow==7.0.0
pipreqs==0.4.10
pyasn1==0.4.8
pycparser==2.20
Pygments==2.6.1
PyMySQL==0.9.3
PyNaCl==1.3.0
pyparsing==2.4.7
Pypubsub==4.0.3
python-dateutil==2.8.1
pytz==2019.3
pywin32==227
PyYAML==3.12
requests==2.18.4
requests-toolbelt==0.9.1
robotframework==3.1.2
robotframework-ride==1.7.4.1
selenium==3.141.0
six==1.14.0
tornado==5.1.1
urllib3==1.22
wxPython==4.0.7.post2
xlrd==1.2.0
xlwt==1.3.0
yarg==0.1.9
